requirements.txt
fastapi==0.104.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
